Revert "Add "com.android.os.statsd" to libprocessgroup apex_available"
This reverts commit bacb44ffb9d402c3cf4f4a866f6defdb5f289fdc.

Reason for revert: Not needed

The original change was made in an attempt to close a perceived hole in
the apex availability checking. However, there is no such hole and so
this change is unnecessary. While there is a dependency path from the
apex to this module one of the dependencies in that path is from a
static library to a shared library. That dependency does not result in
the code that includes the static library being linked to the shared
library at either build time or runtime. Therefore, the shared library
is not part of the APEX and so does not need to be marked as such.

Bug: 152762638
Test: m droid
Change-Id: I77c3b23d28eca99fcb133cb38eb8863277ac7179