Prevent app fuse from indefinitely trying if it gets ENOBUFS
When there is not enough memory to allocate the appfuse socket,
ENOBUFS error could occur repeatly, and the retry will not be ended.
Limit the retry count to prevent infinite repetition.

Limited this to take only 5 seconds which is still quite high but still
better than running indefinitely. We can reduce this further later on.

Test: Manual
Bug: 312503249
Bug: 369519866
Change-Id: I928b7b5e6d0a53831c2b35d48aeb9440e8f6564a