libprocessgroup: Mount cpuset v1 with cpuset_v2_mode
The Android kernel used to carry a patch to restore cpuset masks upon
CPU hotplug onlining, however this was reverted before 6.1. [1] So
devices running kernels >= 6.1 currently suffer from a problem where
cpusets do not include CPUs that they should when CPU hotplugging
occurs.

Fix this by switching from the cpuset filesystem type to cgroup, and by
mounting with the cpuset_v2_mode option which fixes the behavior. It is
possible to use cpuset_v2_mode with the cpuset filesystem type with
kernels that have [2], but deploying this kernel update to already
shipped devices would require waiting 1+ year before we could make this
platform update (and retain the current cpuset filesystem type).

Note that the release_agent option is now intentionally omitted, as it
is unused on Android.

[1] https://r.android.com/2251493
[2] https://lore.kernel.org/all/20250416211752.945849-1-tjmercier@google.com/

Bug: 409240872
Change-Id: I8e39c2831db06d65b6cbbe82f3199272fa85d637