libsnapshot: log message when the WaitForDeviceDelete() fails
Currently when WaitForDeviceDelete() fails, we only log a failure
message without the reason even if snapuserd sent it.

Logging a message will tell us more if snapuserd failed to
do the operation for some reason or did it crash.

This is not fix but a change to get more information in failure path.

Bug: 402491858

Test: th
Change-Id: I30339d983a56133359400e423678d02b9d0f4e4f
Signed-off-by: Sandeep Dhavale <dhavale@google.com>