healthd: BatteryMonitor: gracefully handle errors
When reading battery properties from sysfs, the BatteryMonitor has
historically ignored errors and wrote default values to HealthInfo. This
is generally reasonable because HealthInfo doesn't define values for
"don't know" in most fields, but also results in a bad user experience
if there are transient errors because the battery charge state may be
momentarily reported as 0% which can trigger many user-visible behaviors
including a notification and system theme change if the device
automatically enables battery-saver mode.

This change makes the BatteryMonitor more principled about handling
errors by explicitly checking for errors on every sysfs read, but does
not otherwise change behavior. Future changes may modify the behavior on
error to return more appropriate values.

For cases where success was previously assumed by checking if a read
string's length was positive, care is taken to replicate the old
behavior in case an empty string is read without error. This applies to
non-integer fields and those that default to nonzero values, in most
cases handled by checking for ENOENT to indicate that the relevant
path to search is unknown so other errors still cause the default value
to be output.

Bug: 410996664
Test: on brya, `dumpsys android.hardware.health.IHealth/default` still
      reports correct values and no new messages are logged during
      normal operation.
Test: atest battery_monitor_test
Test: (on brya) atest VtsHalHealthTargetTest
Change-Id: I3b69862ba625ecfff6aa69afa94f614ba11e8f76