Update bootstats to use ms for boot_complete
This change updates bootstats to use milliseconds instead of seconds for
the boot_complete metrics. This aligns the units with the
BootTimeEventElapsedTime atom and provides finer measurements.

Currently, there are no other consumers of this atom. Lab tests that
monitor bootstats directly may identify this CL as a culprit of a boot
time regression, but this change doesn't actually affect boot time.

Test: build, boot, check boot_complete files' mtime
Bug: 427910232
Flag: EXEMPT bugfix
Change-Id: I60b9a88697955a1a795b28716a14940a55e1c40b