Purge unused memory at the end of coldboot
This commit adds a mallopt call with M_PURGE_ALL. This fixes the
regression reported in b/443283645. Previously the uevent handling was
done in subprocesses, purging the memory at the process termination.
However, with the thread pool, we now handle uevents in threads,
accounting the memory consumed by the threads to ueventd. The scudo
allocator in ueventd process may end up not unmapping the unsed memory,
hence we explicitly call mallopt with M_PURGE_ALL to give the unused
memory back to the kernel.

Bug: 443283645
Test: RSS of ueventd decreased
Change-Id: I7571315ee34d2571ce6d85bba618616fdda334db