libsnapshot: add API to query snapuserd operating mode
This change introduces a new API to allow clients to determine the
current operating mode of the snapuserd daemon, which can be either
"ublk" or "dm-user".

The SnapuserdClient now has a GetSnapuserdMode() method that
sends a "get_mode" request to the server. The snapuserd server
handles this request and replies with its currently configured mode.

This provides the necessary mechanism for client code to detect
and make decisions based on the snapuserd configuration.

Bug: 444696988
Test: th

Change-Id: Iaf36b560d72fe67ba8ed34a3ed2e0eacef34cb78
Signed-off-by: Sandeep Dhavale <dhavale@google.com>