libsnapshot: inspect_cow : add merge group tracking to inspect_cow
Adds the ability to track and display merge group numbers in the
inspect_cow tool. The merge group number is calculated based on the
COW header's buffer and block sizes.

This change adds "merge group" which is a set of COW operations that
are processed together by readahead thread and handed over to merge
thread.

The group number changes when the number of operations in the current
group exceeds the calculated maximum which is
(scratch_space_size - metadata)/block_size.
For 2MB scratch space and 8KB metadata it is 510.
Merge group number is only valid for ordered ops (COPY and XOR) as only
those go through readahead thread for special handling as we depend
on data from source device to construct data at the target.

This is useful for debugging and understanding the merging process.
Bug: 448450098
Test: th, run inspect_cow --show_ops <cow.img>

Change-Id: I180e52d7497e4aead8734fbb7b476fcb6b348c3b
Signed-off-by: Sandeep Dhavale <dhavale@google.com>