prefetch: Add metrics collection for prefetch replay
This change introduces a new argument `record_metrics` to prefetch
replay service. When it is enabled prefetch replay collects and reports
metrics. These metrics provide visibility into the effectiveness of the
replay.

The following metrics are now collected:
- `prefetched_records`: The total number of records successfully
  prefetched.
- `prefetched_bytes`: The total number of bytes readed by prefetch.
- `prefetched_pages`: The estimated number of memory pages prefetched.
  If the system's page size cannot be determined, the `prefetched_pages`
  count is reported as "unknown" to avoid providing misleading data.

Upon a successful prefetch replay, performance metrics are written to a
.stat file. This file is created alongside the original .pack file and
shares the same base name. For example, if there's a prefetch.pack file,
a prefetch.stat will be created.

A unit test test_replay_metrics_collection is added to test metric
collection behaviour.

Bug: 464969245
Test: reboot device & found .stat file exsits on /metadata/prefetch
Test: cargo build
Change-Id: Icb240dc051e64d4104b62629d9c6fb2fbe6fd34c